Refactored code and made some methods private
