+++ pkg/driver/deployment/deployer.go
@@ -10,19 +10,30 @@ type FunctionDeployer interface {
 	Clean()
 }
 
-func CreateDeployer(cfg *config.Configuration) FunctionDeployer {
+func CreateDeployer(cfg *config.Configuration) (deployer FunctionDeployer, configuration interface{}) {
 	switch cfg.LoaderConfiguration.Platform {
-	case "AWSLambda", "AWSLambda-RPS":
-		return newAWSLambdaDeployer()
-	case "Dirigent", "Dirigent-RPS", "Dirigent-Dandelion", "Dirigent-Dandelion-RPS":
-		return newDirigentDeployer()
 	case "Knative", "Knative-RPS":
-		return newKnativeDeployer()
+		deployer = &knativeDeployer{}
+		configuration = knativeDeploymentConfiguration{
+			YamlPath:          cfg.YAMLPath,
+			IsPartiallyPanic:  cfg.LoaderConfiguration.IsPartiallyPanic,
+			EndpointPort:      cfg.LoaderConfiguration.EndpointPort,
+			AutoscalingMetric: cfg.LoaderConfiguration.AutoscalingMetric,
+		}
 	case "OpenWhisk", "OpenWhisk-RPS":
-		return newOpenWhiskDeployer()
+		deployer = &openWhiskDeployer{}
+		configuration = openWhiskDeploymentConfiguration{}
+	case "AWSLambda", "AWSLambda-RPS":
+		deployer = &awsLambdaDeployer{}
+		configuration = awsLambdaDeploymentConfiguration{}
+	case "Dirigent", "Dirigent-RPS", "Dirigent-Dandelion", "Dirigent-Dandelion-RPS":
+		deployer = &dirigentDeployer{}
+		configuration = dirigentDeploymentConfiguration{
+			RegistrationServer: cfg.LoaderConfiguration.DirigentControlPlaneIP,
+		}
 	default:
 		logrus.Fatal("Unsupported platform.")
 	}
 
-	return nil
+	return
 }
